+++ src/components/avatar/Avatar.js
@@ -1,40 +1,79 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
+import AvatarAdd from './AvatarAdd';
+import AvatarAnonymous from './AvatarAnonymous';
 import AvatarImage from './AvatarImage';
 import AvatarInitials from './AvatarInitials';
-import AvatarOverlay from './AvatarOverlay';
-import Box, { pickBoxProps, omitBoxProps } from '../box';
+import Box from '../box';
 import cx from 'classnames';
 import theme from './theme.css';
-import { colorsWithout } from '../../constants';
-
-const colors = colorsWithout(['neutral']);
-
-const hashCode = hexString => parseInt(hexString.substr(-5), 16);
-
-const getColor = id => (id ? colors[Math.abs(hashCode(id)) % colors.length] : 'neutral');
+import omit from 'lodash.omit';
 
 class Avatar extends PureComponent {
+  renderComponent = () => {
+    const { creatable, children, editable, imageUrl, fullName, id, onImageChange, size } = this.props;
+
+    if (creatable) {
+      return <AvatarAdd children={children} size={size} />;
+    }
+
+    if (imageUrl) {
+      return (
+        <AvatarImage
+          children={children}
+          editable={editable}
+          image={imageUrl}
+          imageAlt={fullName}
+          onImageChange={onImageChange}
+          size={size}
+        />
+      );
+    }
+
+    if (fullName && id) {
+      return (
+        <AvatarInitials
+          children={children}
+          editable={editable}
+          id={id}
+          name={fullName}
+          onImageChange={onImageChange}
+          size={size}
+        />
+      );
+    }
+
+    return <AvatarAnonymous children={children} size={size} />;
+  };
+
   render() {
-    const { className, editable, imageUrl, fullName, id, onImageChange, size, shape, ...others } = this.props;
+    const { className, size, shape, ...others } = this.props;
     const avatarClassNames = cx(theme['avatar'], theme[`is-${size}`], theme[`is-${shape}`], className);
-    const boxProps = pickBoxProps(others);
-    const propsWithoutBoxProps = omitBoxProps(others);
+    const restProps = omit(others, [
+      'children',
+      'creatable',
+      'editable',
+      'imageUrl',
+      'fullName',
+      'id',
+      'onImageChange',
+    ]);
 
     return (
-      <Box className={avatarClassNames} {...boxProps}>
-        {imageUrl ? (
-          <AvatarImage image={imageUrl} imageAlt={fullName} {...propsWithoutBoxProps} />
-        ) : (
-          <AvatarInitials color={getColor(id)} name={fullName} {...propsWithoutBoxProps} />
-        )}
-        {editable && (size === 'large' || size === 'hero') && <AvatarOverlay onClick={onImageChange} />}
+      <Box className={avatarClassNames} {...restProps}>
+        {this.renderComponent()}
       </Box>
     );
   }
 }
 
 Avatar.propTypes = {
+  /** Component that will be placed top right of the avatar image. */
+  children: PropTypes.any,
+  /** A class name for the wrapper to give custom styles. */
+  className: PropTypes.string,
+  /** If true, an add user icon will be shown. */
+  creatable: PropTypes.bool,
   /** If true, an overlay will be shown with edit icon. */
   editable: PropTypes.bool,
   /** The image url to show as an avatar. */
@@ -52,6 +91,8 @@ Avatar.propTypes = {
 };
 
 Avatar.defaultProps = {
+  creatable: false,
+  editable: false,
   shape: 'circle',
   size: 'medium',
 };