+++ src/components/avatar/AvatarImage.js
@@ -1,15 +1,16 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import cx from 'classnames';
 import theme from './theme.css';
+import AvatarOverlay from './AvatarOverlay';
 
 class AvatarImage extends PureComponent {
   render() {
-    const { children, image, imageAlt, imageClassName, ...others } = this.props;
+    const { children, editable, image, imageAlt, onImageChange, size } = this.props;
 
     return (
-      <div className={theme['avatar-image']} {...others} data-teamleader-ui="avatar-image">
-        <img alt={imageAlt} src={image} className={cx(theme['image'], imageClassName)} />
+      <div className={theme['avatar-image']} data-teamleader-ui="avatar-image">
+        <img alt={imageAlt} src={image} className={theme['image']} />
+        {editable && (size === 'large' || size === 'hero') && <AvatarOverlay onClick={onImageChange} />}
         {children && <div className={theme['children']}>{children}</div>}
       </div>
     );
@@ -19,12 +20,16 @@ class AvatarImage extends PureComponent {
 AvatarImage.propTypes = {
   /** Component that will be placed top right of the avatar image. */
   children: PropTypes.any,
+  /** If true, an overlay will be shown with edit icon. */
+  editable: PropTypes.bool,
   /** An image source or an image element. */
   image: PropTypes.string,
   /** An alternative text for the image element. */
   imageAlt: PropTypes.string,
-  /** A class name for the image to give custom styles. */
-  imageClassName: PropTypes.string,
+  /** Callback function that is fired when user clicks the edit icon. */
+  onImageChange: PropTypes.func,
+  /** The size of the avatar. */
+  size: PropTypes.oneOf(['tiny', 'small', 'medium', 'large', 'hero']),
 };
 
 export default AvatarImage;