+++ src/components/avatar/AvatarInitials.js
@@ -1,10 +1,25 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import cx from 'classnames';
 import theme from './theme.css';
+import AvatarOverlay from './AvatarOverlay';
+import Box from '../box';
 import { Heading4 } from '../typography';
+import { colorsWithout } from '../../constants';
+
+const colors = colorsWithout(['neutral']);
+const hashCode = hexString => parseInt(hexString.substr(-5), 16);
 
 class AvatarInitials extends PureComponent {
+  getColor = () => {
+    const { id } = this.props;
+
+    if (!id) {
+      return 'neutral';
+    }
+
+    return colors[Math.abs(hashCode(id)) % colors.length];
+  };
+
   getInitials = () => {
     const { name } = this.props;
     const splitName = name.split(' ');
@@ -17,14 +32,19 @@ class AvatarInitials extends PureComponent {
   };
 
   render() {
-    const { children, color, shape, ...others } = this.props;
-    const avatarClassNames = cx(theme['avatar-initials'], theme[color]);
+    const { children, editable, onImageChange, size } = this.props;
 
     return (
-      <div className={avatarClassNames} {...others} data-teamleader-ui="avatar-initials">
+      <Box
+        backgroundColor={this.getColor()}
+        backgroundTint="normal"
+        className={theme['avatar-initials']}
+        data-teamleader-ui="avatar-initials"
+      >
         <Heading4 className={theme['initials']}>{this.getInitials()}</Heading4>
+        {editable && (size === 'large' || size === 'hero') && <AvatarOverlay onClick={onImageChange} />}
         {children && <div className={theme['children']}>{children}</div>}
-      </div>
+      </Box>
     );
   }
 }
@@ -32,14 +52,13 @@ class AvatarInitials extends PureComponent {
 AvatarInitials.propTypes = {
   /** Component that will be placed top right of the avatar image. */
   children: PropTypes.any,
-  /** The color of the avatar. */
-  color: PropTypes.oneOf(['neutral', 'mint', 'aqua', 'violet', 'teal', 'gold', 'ruby']),
+  /** The uuid to determine the background color of the avatar. */
+  id: PropTypes.string,
   /** The name for in the avatar. */
   name: PropTypes.string,
 };
 
 AvatarInitials.defaultProps = {
-  color: 'neutral',
   name: 'Michael Scott',
 };